deepen: dissolve email-core into the email module entry point
`lib/email/email-core.ts` (39 LOC) existed solely to construct the
default `ResendClient` singleton, declare a one-line `sendEmail(message,
options)` wrapper around `emailClient.sendEmail(message, options)`, and
re-export the same names. `lib/email/index.ts` then re-imported those
three names from email-core and re-exported them — every external caller
reached `emailClient`/`sendEmail` via the index, so email-core was a
pure intermediate.

Deletion test: deleting `email-core.ts` and moving the singleton +
wrapper into `lib/email/index.ts` does not concentrate complexity — the
33 lines of substance are now the public interface of `@/lib/email`
directly. Two direct importers (`lib/cron/handlers/weekly-digest-handler.ts`
and `lib/email/trial-emails.ts`) migrate to `'@/lib/email'` /
`'./index'`; every other caller already imported from the index and is
unchanged.

The interface is unchanged: `emailClient`, `sendEmail`, `ResendClient`,
`resendConfig`, `NotificationPreference`, and the types from
`./types` all remain exported under the same names from `@/lib/email`.

Note: the second `resendClient` singleton in `resend-client.ts` (line
632) is left in place because `lib/monitoring/alert-service.ts` imports
it directly and `jest.setup.js` only mocks the `ResendClient` class, not
the singleton. Collapsing the two singletons is a follow-up that
requires updating the mock surface.
